+++ project/ProjectPlugin.scala
@@ -55,13 +55,6 @@ object ProjectPlugin extends AutoPlugin {
         %%("monocle-core", V.monocle),
         %%("fs2-reactive-streams", V.fs2ReactiveStreams),
         %%("fs2-core", V.fs2),
-
-        %%("http4s-dsl", V.http4s),
-        %%("http4s-blaze-server", V.http4s),
-        %%("http4s-circe", V.http4s),
-        %%("circe-generic"),
-        %%("http4s-blaze-client", V.http4s),
-
         %%("pbdirect", V.pbdirect),
         %%("avro4s", V.avro4s),
         %%("log4s", V.log4s),
@@ -205,7 +198,7 @@ object ProjectPlugin extends AutoPlugin {
       )
     )
 
-    lazy val rpcHttpServerSettings: Seq[Def.Setting[_]] = Seq(
+    lazy val rpcHttpSettings: Seq[Def.Setting[_]] = Seq(
       libraryDependencies ++= Seq(
         %%("http4s-dsl", V.http4s),
         %%("http4s-blaze-server", V.http4s),
